My desktop tensorflow

+++ tensorflow/core/framework/function.h
@@ -18,14 +18,12 @@ limitations under the License.
 
 #include <unordered_map>
 
-#include <vector>
 #include "tensorflow/core/framework/attr_value_util.h"
 #include "tensorflow/core/framework/function.pb.h"
 #include "tensorflow/core/framework/graph.pb.h"
 #include "tensorflow/core/framework/node_def_util.h"
 #include "tensorflow/core/framework/op.h"
 #include "tensorflow/core/framework/types.h"
-#include "tensorflow/core/platform/macros.h"
 #include "tensorflow/core/platform/protobuf.h"
 
 namespace tensorflow {
@@ -83,7 +81,7 @@ class FunctionDefHelper {
     return FunctionRef(name, {});
   }
 
-  // Node is used to construct FunctionDef.Node using initialization
+  // Node is used to consturct FunctionDef.Node using initialization
   // lists. E.g.,
   //  Node n = {{"z"}, "Mul", {"x", "y"}, {{"T", "$T"}}};  // z = x * y
   struct Node {
@@ -127,7 +125,7 @@ class FunctionDefHelper {
     n.attr.push_back({"dtype", dtype});
     int64 num = vals.size();
     Tensor t(dtype, TensorShape({num}));
-    for (size_t i = 0; i < vals.size(); ++i) {
+    for (int i = 0; i < vals.size(); ++i) {
       t.flat<T>()(i) = vals[i];
     }
     n.attr.push_back({"value", t});
